+++ src/main/java/roomescape/dao/ReservationMemoryDao.java
@@ -8,7 +8,7 @@ import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationRequest;
 
-public class ReservationMemoryDao implements ReservationRepository{
+public class ReservationMemoryDao implements ReservationDao {
     private final List<Reservation> reservations = new ArrayList<>();
     private final AtomicLong index = new AtomicLong(1);
