+++ src/main/java/roomescape/service/ReservationService.java
@@ -2,7 +2,7 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.dao.ReservationRepository;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
@@ -12,22 +12,22 @@ import roomescape.dto.ReservationRequest;
 public class ReservationService {
 
     private final ReservationTimeDao reservationTimeDao;
-    private final ReservationRepository reservationRepository;
-    public ReservationService(ReservationTimeDao reservationTimeDao, ReservationRepository reservationRepository) {
-        this.reservationRepository = reservationRepository;
+    private final ReservationDao reservationDao;
+    public ReservationService(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
+        this.reservationDao = reservationDao;
         this.reservationTimeDao = reservationTimeDao;
     }
 
     public List<Reservation> selectReservations() {
-        return reservationRepository.selectReservations();
+        return reservationDao.selectReservations();
     }
 
     public Reservation createReservation(ReservationRequest reservationRequest) {
         ReservationTime reservationTime = reservationTimeDao.selectById(reservationRequest.getTimeId());
-        return reservationRepository.insertReservation(reservationRequest, reservationTime);
+        return reservationDao.insertReservation(reservationRequest, reservationTime);
     }
 
     public void deleteReservation(long id) {
-        reservationRepository.deleteReservation(id);
+        reservationDao.deleteReservation(id);
     }
 }
